fixing search ordering
